+++ oidc-ui/package.json
@@ -47,6 +47,7 @@
   "devDependencies": {
     "autoprefixer": "^10.4.8",
     "postcss": "^8.4.16",
-    "tailwindcss": "^3.1.8"
+    "tailwindcss": "^3.1.8",
+    "tailwindcss-dir": "^4.0.0"
   }
 }
