+++ oidc-ui/public/locales/ar.json
@@ -1,0 +1,133 @@
+{
+  "langCode": "ar",
+  "background": {
+    "home": "مسكن",
+    "polices": "شرطة",
+    "departments": "الإدارات",
+    "gov_circulars": "التعاميم الحكومية",
+    "recruitment": "توظيف",
+    "about_us": "معلومات عنا",
+    "contact_us": "اتصل بنا",
+    "health_portal": "اتصل بنا",
+    "confirmed": "مؤكد",
+    "recovered": "تعافى",
+    "active": "نشيط",
+    "registration": "تسجيل",
+    "language": "لغة"
+  },
+  "login": {
+    "sign_in_with_health_portal": "قم بتسجيل الدخول إلى بوابة الصحة",
+    "username": "اسم المستخدم",
+    "password": "كلمة المرور",
+    "submit": "إرسال",
+    "or": "أو",
+    "sign_in_with_mosip": "تسجيل الدخول باستخدام MOSIP",
+    "dont_have_existing_account": "ليس لديك حساب موجود؟",
+    "sign_up_here": "سجل هنا"
+  },
+  "signup": {
+    "sign_up_with_health_portal": "قم بالتسجيل في بوابة الصحة",
+    "first_name": "الاسم الاول",
+    "last_name": "الكنية",
+    "password": "كلمة المرور",
+    "submit": "إرسال",
+    "or": "أو",
+    "sign_up_with_mosip": "قم بالتسجيل مع MOSIP",
+    "already_have_account": "هل لديك حساب؟",
+    "sign_in_here": "تسجيل الدخول هنا"
+  },
+  "userprofile": {
+    "welcome_msg": "مرحبا بكم في الخدمات الصحية",
+    "loading_msg": "جار التحميل! أرجو الإنتظار...",
+    "profile_picture": "الصوره الشخصيه",
+    "email_address": "عنوان البريد الالكترونى",
+    "gender": "جنس",
+    "phone_number": "رقم الهاتف",
+    "birth_date": "تاريخ الميلاد",
+    "address": "تبوك",
+    "show_raw_user_info": "إظهار معلومات المستخدم الأولية",
+    "hide_raw_user_info": "إخفاء معلومات المستخدم الأولية"
+  },
+  "registration": {
+    "child_registration": "تسجيل الطفل",
+    "full_name": "الاسم الكامل",
+    "gender": "جنس تذكير أو تأنيث",
+    "date_of_birth": "تاريخ الولادة",
+    "city": "مدينة",
+    "address": "تبوك",
+    "fetch_details": "إحضار التفاصيل من MOSIP",
+    "register": "يسجل",
+    "loading_msg": "تحميل! يرجى الانتظار...",
+    "phone_number": "رقم التليفون"
+  },
+  "error": {
+    "sign_in_failed": "المصادقة فشلت! حاول تسجيل الدخول باستخدام MOSIP",
+    "sign_up_failed": "فشل انشاء الحساب! حاول تسجيل الدخول باستخدام MOSIP",
+    "invalid_details": "تفاصيل غير صحيحة! حاول جلب التفاصيل من MOSIP",
+    "authCode_missing": "AuthCode مفقود",
+    "no_response_msg": "لا يوجد رد",
+    "no_devices_found_msg": "لم يتم العثور علي الجهاز. يرجى التحقق من الاتصال وإعادة المحاولة.",
+    "link_code_refresh_failed": "فشل في التحميل",
+    "link_code_status_failed": "فشل في التحميل",
+    "failed_to_link": "فشل الارتباط",
+    "qr_code_expired": "انتهت صلاحية رمز QRCode",
+    "authorization_failed": "فشل التفويض",
+    "device_disc_failed": "فشل اكتشاف الجهاز",
+    "device_not_found_msg": "لم يتم العثور علي الجهاز",
+    "biometric_capture_failed_msg": "فشل التقاط القياسات الحيوية:",
+    "authentication_failed_msg": "المصادقة فشلت",
+    "send_otp_failed_msg": "فشل إرسال OTP",
+    "parsing_error_msg": "غير قادر على تحليل",
+    "Network Error": "خطأ في الشبكة",
+    "invalid_request": "طلب غير صالح",
+    "invalid_client_id": "معرف العميل غير صالح",
+    "invalid_client_name": "اسم العميل غير صالح",
+    "invalid_uri": "URL غير صالح",
+    "invalid_rp_id": "معرّف جهة الاعتماد غير صالحة",
+    "invalid_response_type": "نوع الرد غير صالح / غير مدعوم",
+    "invalid_grant_type": "نوع المنحة غير صالح / غير مدعوم",
+    "invalid_scope": "نطاق غير صالح / غير مدعوم",
+    "invalid_redirect_uri": "عنوان URL لإعادة التوجيه غير صالح",
+    "invalid_display": "عرض غير صالح / غير مدعوم",
+    "invalid_prompt": "موجه واجهة مستخدم غير صالح / غير مدعوم",
+    "invalid_assertion_type": "نوع تأكيد العميل غير صالح / غير معتمد",
+    "invalid_client_auth": "أنواع مصادقة العميل غير صالحة / غير مدعومة",
+    "invalid_transaction": "معاملة أو معاملة غير صالحة منتهية الصلاحية",
+    "invalid_assertion": "تأكيد العميل غير صالح",
+    "invalid_acr": "مراجع فئة سياق المصادقة غير صالحة / غير مدعومة",
+    "invalid_token": "رمز الوصول غير صالح / منتهي الصلاحية",
+    "auth_failed": "المصادقة فشلت",
+    "auth_passed": "نجاح المصادقة",
+    "acr_amr_mapping_not_found": "لم يتم العثور على ملف تعيين ACR-AMR",
+    "no_acr_registered": "لم يتم تسجيل مراجع فئة سياق المصادقة (ACR)",
+    "invalid_algorithm": "خوارزمية تجزئة غير صالحة / غير مدعومة",
+    "unknown_error": "خطأ غير معروف",
+    "invalid_input": "مدخل غير صالح",
+    "duplicate_client_id": "معرّف عميل مكرر",
+    "invalid_claim": "مطالبات مستخدم غير صالحة / غير مدعومة",
+    "invalid_public_key": "مفتاح العميل العمومي غير صالح",
+    "send_otp_failed": "فشل send-otp. معرف فردي غير صالح / غير معروف.",
+    "missing_header": "العنوان غير موجود برمز الحامل المميز في الطلب: التفويض.",
+    "invalid_permitted_scope": "نطاق تفويض غير صالح مسموح به",
+    "invalid_identifier": "تم تقديم معرّف فردي غير صالح.",
+    "invalid_otp_channel": "تم توفير قناة OTP غير صالحة.",
+    "invalid_no_of_challenges": "اختبارات المصادقة الفارغة أو الفارغة غير مسموح بها.",
+    "invalid_transaction_id": "معاملة فارغة أو فارغة لم يُسمح بها.",
+    "invalid_auth_factor_type": "نوع عامل المصادقة الفارغ أو الفارغ غير مسموح به.",
+    "invalid_challenge": "مطلوب تحدي المصادقة.",
+    "invalid_status": "تم تقديم قيمة حالة غير مدعومة.",
+    "invalid_accepted_claim": "تم قبول مطالبات المستخدم غير الصالحة",
+    "invalid_auth_code": "تم تقديم رمز تفويض غير صالح.",
+    "auth_factor_mismatch": "لا تتطابق عوامل المصادقة المقدمة مع عوامل المصادقة التي تم حلها.",
+    "unsupported_id_format": "تنسيق المعرف غير مدعوم.",
+    "link_code_gen_failed": "فشل إنشاء رمز الارتباط. حاول مرة أخرى.",
+    "invalid_link_code": "كود الارتباط غير صالح.",
+    "transaction_not_found": "لم يتم العثور على الصفقة.",
+    "transaction_already_linked": "الصفقة مرتبطة بالفعل.",
+    "subscription_failed": "فشل الاشتراك في موضوع كفى.",
+    "response_timeout": "انتهت مهلة الرد المؤجل.",
+    "invalid_state": "وجدت المعاملة في حالة غير صالحة.",
+    "authorization_failed_msg": "فشل التفويض",
+    "consent_request_rejected": "تم رفض طلب الموافقة"
+  }
+}
