+++ oidc-ui/public/locales/en.json
@@ -12,7 +12,8 @@
     "confirmed": "Confirmed",
     "recovered": "Recovered",
     "active": "Active",
-    "registration": "Registration"
+    "registration": "Registration",
+    "language": "Language"
   },
   "login": {
     "sign_in_with_health_portal": "Sign In to the Health portal",
