+++ oidc-ui/src/pages/Login.js
@@ -2,15 +2,14 @@ import React from "react";
 import Background from "../components/Background";
 import Login from "../components/Login";
 import clientService from "../services/clientService";
-import langConfigService from "../services/langConfigService";
 
-export default function LoginPage() {
+export default function LoginPage({ langOptions }) {
   return (
     <Background
-      langConfigService={langConfigService}
       component={React.createElement(Login, {
         clientService: clientService,
       })}
+      langOptions={langOptions}
     />
   );
 }
