+++ oidc-ui/src/pages/SignUp.js
@@ -2,15 +2,14 @@ import React from "react";
 import Background from "../components/Background";
 import SignUp from "../components/SignUp";
 import clientService from "../services/clientService";
-import langConfigService from "../services/langConfigService";
 
-export default function SignUpPage() {
+export default function SignUpPage({ langOptions }) {
   return (
     <Background
-      langConfigService={langConfigService}
       component={React.createElement(SignUp, {
         clientService: clientService,
       })}
+      langOptions={langOptions}
     />
   );
 }
