+++ oidc-ui/src/pages/UserProfile.js
@@ -1,16 +1,15 @@
 import React from "react";
 import Background from "../components/Background";
 import UserProfile from "../components/UserProfile";
-import langConfigService from "../services/langConfigService";
 import oidcService from "../services/oidcService";
 
-export default function UserProfilePage() {
+export default function UserProfilePage({ langOptions }) {
   return (
     <Background
-      langConfigService={langConfigService}
       component={React.createElement(UserProfile, {
         oidcService: oidcService,
       })}
+      langOptions={langOptions}
     />
   );
 }
