+++ oidc-ui/tailwind.config.js
@@ -5,5 +5,5 @@ module.exports = {
   theme: {
     extend: {},
   },
-  plugins: [],
+  plugins: [require('tailwindcss-dir')()],
 }
